backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1
scipy==1.11.4
plotly==5.18.0
kaleido==0.2.1
numpy==1.26.3
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.3
